flaskAPI initially setup
