+++ data/elastic.py
@@ -7,15 +7,15 @@ import time
 
 
 # Password for the 'elastic' user generated by Elasticsearch
-ELASTIC_PASSWORD = "<REPLACE>"
+ELASTIC_PASSWORD = "lWwjJoVHt50J00-WaYzm"
 
 # CERT_FINGERPRINT
-CERT_FINGERPRINT = "<REPLACE>"
+CERT_FINGERPRINT = "B7:F7:50:A9:18:D7:E4:B9:AD:48:F9:82:93:74:09:39:23:45:67:C7:6C:5C:3C:6B:35:B9:DA:4C:C4:B1:7F:86"
 
 # Create the client instance
 client = Elasticsearch(
     "https://localhost:9200",
-    ssl_assert_fingerprint=CERT_FINGERPRINT,
+    ssl_assert_fingerprint=(CERT_FINGERPRINT),
     basic_auth=("elastic", ELASTIC_PASSWORD)
 )
