+++ data/flaskAPI.py
@@ -1,0 +1,58 @@
+import pandas as pd
+import sys
+import names
+import numpy as np
+from flask import Flask, jsonify, make_response
+from flask_restful import Resource, Api
+import ast
+import json
+from elasticsearch import Elasticsearch
+
+
+# Password for the 'elastic' user generated by Elasticsearch
+ELASTIC_PASSWORD = "lWwjJoVHt50J00-WaYzm"
+
+# CERT_FINGERPRINT
+CERT_FINGERPRINT = "B7:F7:50:A9:18:D7:E4:B9:AD:48:F9:82:93:74:09:39:23:45:67:C7:6C:5C:3C:6B:35:B9:DA:4C:C4:B1:7F:86"
+
+# Create the es instance
+es = Elasticsearch(
+    "https://localhost:9200",
+    ssl_assert_fingerprint=(CERT_FINGERPRINT),
+    basic_auth=("elastic", ELASTIC_PASSWORD)
+)
+
+
+class Users(Resource):
+    
+    def get(self):                
+        es.indices.refresh(index="dating_profiles")
+        search_param = {
+            "query": {
+                "fields": {
+                    "_id": [1]
+                }
+            }
+        }
+        resp = es.get(index="dating_profiles", id=np.random.choice(100))
+        print(resp)
+        print("json dumps")
+        print(json.dumps(resp['_source']))
+        
+        # jso = json.dumps(resp, indent=4)
+        # print(jso)
+        return json.dumps(resp['_source'])
+    # methods go here
+    pass
+    
+class Locations(Resource):
+    # methods go here
+    pass
+
+app = Flask(__name__)
+api = Api(app)
+
+api.add_resource(Users, '/get_user')  # '/users' is our entry point for Users
+api.add_resource(Locations, '/add_user')  # and '/locations' is our entry point for Locations
+if __name__ == '__main__':
+    app.run(debug=True)
