+++ data/process.py
@@ -1,0 +1,14 @@
+import pandas as pd
+import sys
+import names
+import numpy as np
+import json
+import ast
+
+f = open('data.json')
+data = json.load(f)
+
+# hypothetically somehow extract orientation + gender, + age
+
+print(data)
+
